Fixed: don't try to delete series folder if it doesn't exist

+++ src/NzbDrone.Core/MediaFiles/RecycleBinProvider.cs
@@ -170,7 +170,10 @@ namespace NzbDrone.Core.MediaFiles
         {
             if (message.DeleteFiles)
             {
-                DeleteFolder(message.Series.Path);
+                if (_diskProvider.FolderExists(message.Series.Path))
+                {
+                    DeleteFolder(message.Series.Path);
+                }
             }
         }
